feat: 策略模式优惠折扣计算

+++ grouph-buy-market-study-infrastructure/src/main/java/com/hjs/study/infrastructure/dao/po/GroupBuyDiscount.java
@@ -7,11 +7,6 @@ import lombok.NoArgsConstructor;
 
 import java.util.Date;
 
-/**
- * @author Fuzhengwei bugstack.cn @小傅哥
- * @description 折扣配置
- * @create 2024-12-07 10:06
- */
 @Data
 @Builder
 @AllArgsConstructor
@@ -41,7 +36,7 @@ public class GroupBuyDiscount {
     /**
      * 折扣类型（0:base、1:tag）
      */
-    private Byte discountType;
+    private Integer discountType;
 
     /**
      * 营销优惠计划（ZJ:直减、MJ:满减、N元购）
